Edge/requirements.txt
paho-mqtt==1.6.1
pillow==9.2.0
tensorflow==2.10.0